Use a navigation router
