+++ package.json
@@ -9,6 +9,7 @@
 	"dependencies": {
 		"react": "16.0.0",
 		"react-native": "0.50.3",
+		"react-navigation": "1.0.0-beta.19",
 		"react-primitives": "0.4.4"
 	},
 	"devDependencies": {
