+++ src/App.js
@@ -1,26 +1,14 @@
 import React, { Component } from 'react';
-import { StyleSheet, Text, View } from 'react-primitives';
+import { StackNavigator } from 'react-navigation';
+
+import HomeScreen from './routes/HomeScreen';
+
+const Main = StackNavigator({
+    Home: { screen: HomeScreen },
+});
 
 export default class App extends Component {
     render() {
-        return (
-            <View style={styles.container}>
-                <Text style={styles.welcome}>Welcome to React Native!</Text>
-            </View>
-        );
+        return <Main />;
     }
 }
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-        backgroundColor: '#F5FCFF',
-    },
-    welcome: {
-        fontSize: 20,
-        textAlign: 'center',
-        margin: 10,
-    },
-});
